+++ lib/src/pages/tags/tag_manager_create_dialog.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../core/tag_rules.dart';
 import '../../models/platform_models.dart';
+import '../../widgets/maintenance_feedback.dart';
 
 // ignore_for_file: slash_for_doc_comments
 
@@ -21,7 +22,7 @@ Widget createTagDialogSmokeHarness() {
     home: Scaffold(
       body: Builder(
         builder: (context) => FilledButton(
-          onPressed: () => showDialog<void>(
+          onPressed: () => showMaintenanceDialog<void>(
             context: context,
             builder: (_) => const CreateTagDialog(groups: <TagGroup>[]),
           ),